services/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.4
django-cors-headers==3.11.0
djangorestframework==3.13.1
joblib==1.1.0
mysqlclient==2.1.0
numpy==1.22.3
pytz==2022.1
scikit-learn==1.0.2
scipy==1.8.0
sklearn==0.0
sqlparse==0.4.2
threadpoolctl==3.1.0
tzdata==2022.1
